+++ src/playground-project.ts
@@ -179,11 +179,49 @@ export class PlaygroundProject extends LitElement {
   }
 
   /**
+   * A pristine copy of the original project files, used for the `modified`
+   * getter.
+   */
+  private _pristineFiles?: SampleFile[];
+
+  /**
+   * Cached value for the `modified` getter. When undefined, the modified state
+   * is unknown and must be computed.
+   */
+  private _modified: boolean | undefined = false;
+
+  /**
+   * Indicates whether the user has modified, added, or removed any project
+   * files. Resets whenever a new project is loaded.
+   */
+  get modified(): boolean {
+    if (this._modified === undefined) {
+      if (this._files === undefined && this._pristineFiles === undefined) {
+        this._modified = false;
+      } else if (
+        this._files === undefined ||
+        this._pristineFiles === undefined
+      ) {
+        this._modified = true;
+      } else {
+        this._modified = !playgroundFilesDeepEqual(
+          this._files,
+          this._pristineFiles
+        );
+      }
+    }
+    return this._modified;
+  }
+
+  /**
    * A unique identifier for this instance so the service worker can keep an
    * independent cache of files for it.
    */
   private readonly _sessionId: string = generateUniqueSessionId();
 
+  /**
+   * The active project files.
+   */
   private _files?: SampleFile[];
 
   @internalProperty()
@@ -313,6 +351,9 @@ export class PlaygroundProject extends LitElement {
         source as void; // Exhaustive check.
         break;
     }
+    this._pristineFiles =
+      this._files && JSON.parse(JSON.stringify(this._files));
+    this._modified = false;
     this.dispatchEvent(new CustomEvent('filesChanged'));
     this.save();
   }
@@ -543,6 +584,15 @@ export class PlaygroundProject extends LitElement {
     return true;
   }
 
+  editFile(file: SampleFile, newContent: string) {
+    // Note this method takes the file object itself rather than the name like
+    // add/delete/rename, because edits happen at high frequency so we don't
+    // want to be doing any searches.
+    file.content = newContent;
+    this._modified = undefined;
+    this.saveDebounced();
+  }
+
   addFile(name: string) {
     if (!this._files || !this.isValidNewFilename(name)) {
       return;
@@ -559,13 +609,22 @@ export class PlaygroundProject extends LitElement {
         contentType: typeFromFilename(name),
       });
     }
+    this._modified = undefined;
     this.requestUpdate();
     this.dispatchEvent(new CustomEvent('filesChanged'));
     this.save();
   }
 
   deleteFile(filename: string) {
-    this._files = this._files?.filter((file) => file.name !== filename);
+    if (!this._files) {
+      return;
+    }
+    const idx = this._files.findIndex((file) => file.name === filename);
+    if (idx < 0) {
+      return;
+    }
+    this._files = [...this._files.slice(0, idx), ...this._files.slice(idx + 1)];
+    this._modified = undefined;
     this.dispatchEvent(new CustomEvent('filesChanged'));
     this.save();
   }
@@ -585,6 +644,7 @@ export class PlaygroundProject extends LitElement {
     file.name = newName;
     file.contentType = typeFromFilename(newName);
     this._files = [...this._files];
+    this._modified = undefined;
     this.dispatchEvent(new CustomEvent('filesChanged'));
     this.save();
   }
@@ -837,4 +897,36 @@ const outdent = (str: string): string => {
     }
   }
   return str.replace(RegExp(`^\\s{${shortestIndent ?? 0}}`, 'gm'), '');
+};
+
+/**
+ * Test whether two lists of Playground files are deeply equal.
+ */
+const playgroundFilesDeepEqual = (
+  filesA: SampleFile[],
+  filesB: SampleFile[]
+): boolean => {
+  if (filesA.length !== filesB.length) {
+    return false;
+  }
+  for (let i = 0; i < filesA.length; i++) {
+    const fileA = filesA[i];
+    const fileB = filesB[i];
+    if (
+      fileA.name !== fileB.name ||
+      fileA.contentType !== fileB.contentType ||
+      fileA.hidden !== fileB.hidden ||
+      fileA.label !== fileB.label
+    ) {
+      return false;
+    }
+  }
+  for (let i = 0; i < filesA.length; i++) {
+    const fileA = filesA[i];
+    const fileB = filesB[i];
+    if (fileA.content !== fileB.content) {
+      return false;
+    }
+  }
+  return true;
 };