+++ CHANGELOG.md
@@ -27,6 +27,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   The `module`, `import`, and `development`
   [conditions](https://nodejs.org/api/packages.html#packages_conditional_exports)
   are enabled.
+  
+### Added
+
+- Added `modified` property to `<playground-project>` and `<playground-ide>`
+  which indicates whether the user has modified, added, or removed any project
+  files. Resets whenever a new project is loaded.
 
 ## [0.12.1] - 2021-08-26